# server/src/openapi/openapi.yaml
...
openapi: 3.0.3
info:
  #  title: Mentor Success API
  description: API for the Mentor Success application
  contact:
    email: whitney@aiontechnology.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.1.0
servers:
  - url: "https://lj6z99n04d.execute-api.us-west-2.amazonaws.com/mentorsuccess-test-service-deployment"
...
paths:
...
  /api/v1/schools/{schoolId}/invitations:
    parameters:
      - in: header
        name: Authorization
        schema:
          type: string
      - in: header
        name: Accept
        schema:
          type:string
      - in: path
        name: schoolId
        schema:
          type: string
          format: uuid
        required: true
        description: The id of the desired school

    # OPTIONS for /api/v1/schools/{schoolId}/invitations
    options:
      tags:
        - school
        - invitation
      responses:
        200:
          description: Success
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: "string"
            Access-Control-Allow-Methods:
              schema:
                type: "string"
            Access-Control-Allow-Headers:
              schema:
                type: "string"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
      x-amazon-apigateway-auth:
        type: None
      x-amazon-apigateway-integration:
        responses:
          default:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Methods: "'DELETE,GET,HEAD,OPTIONS,PATCH,POST,PUT'"
              method.response.header.Access-Control-Allow-Headers: "'Content-Type,Authorization,X-Amz-Date,X-Api-Key,X-Amz-Security-Token'"
              method.response.header.Access-Control-Allow-Origin: "'*'"
        requestTemplates:
          application/json: "{\"statusCode\": 200}"
        passthroughBehavior: "when_no_match"
        type: "mock"

    # POST for /api/v1/schools/{schoolId}/invitations
    post:
      summary: Create a student invitation
      description: Invite a student to join the program
      tags:
        - school
        - invitation
      operationId: createInvitation
      requestBody:
        description: The invitation
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvitationModel'
      responses:
        201:
          description: The invitation is created
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationResultModel'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/NotAuthenticated'
        403:
          $ref: '#/components/responses/NotAuthorized'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/Error'
      x-amazon-apigateway-auth:
        type: None
      x-amazon-apigateway-integration:
        uri: "${lb_url}/api/v1/schools/{schoolId}/invitations"
        responses:
          201:
            statusCode: "201"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          400:
            statusCode: "400"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          401:
            statusCode: "401"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          403:
            statusCode: "403"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          404:
            statusCode: "404"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          default:
            statusCode: "500"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
        requestParameters:
          integration.request.header.Authorization: "method.request.header.Authorization"
          integration.request.header.X-Forwarded-Proto: "'https'"
          integration.request.path.schoolId: "method.request.path.schoolId"
          integration.request.header.X-Forwarded-Host: "stageVariables.api_url"
        passthroughBehavior: "when_no_match"
        connectionType: "VPC_LINK"
        connectionId: "$${stageVariables.vpc_link_id}"
        httpMethod: "POST"
        type: "http"

  # /api/v1/schools/{schoolId}/mentors ---------------------------------------------
...
components:
  responses:
    BadRequest:
      description: There was an error while creating the person
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BadRequestModel'
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
    Error:
      description: Encountered a server error
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
    NotAuthenticated:
      description: Access token is missing or invalid
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
    NotAuthorized:
      description: User does not have permissions for operation
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
    NotFound:
      description: Requested resource could not be found
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
...
  schemas:
...
    BadRequestModel:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: string
        error:
          type: object
        message:
          type: string
        path:
          type: string

    # BehaviorModelCollection ------------------------------------------------------
...
    Empty:
      type: object

    # GameModel --------------------------------------------------------------------
...
    InvitationModel:
      type: object
      properties:
        parentFirstName:
          type: string
          maxLength: 100
          example: Pam
        parentLastName:
          type: string
          maxLength: 100
          example: Parent
        parentEmailAddress:
          type: string
          maxLength: 100
          example: pam@yahoo.com
        studentFirstName:
          type: string
          maxLength: 100
          example: Sally
        studentLastName:
          type: string
          maxLength: 100
          example: Student

    # InvitationResultModel --------------------------------------------------------
    InvitationResultModel:
      type: object
      allOf:
        - $ref: '#/components/schemas/InvitationModel'

    # LeadershipSkillModelCollection -----------------------------------------------
...